engine/texture: fixed hotspot location for mirrored textures

+++ src/engine/texture.h
@@ -27,7 +27,7 @@ one sprite included in the "texture atlas" has.
 */
 struct subtexture {
 	/** x,y starting coordinates and width/height of the subtexture */
-	unsigned int x, y, w, h;
+	float x, y, w, h;
 
 	/** hotspot coordinates. */
 	unsigned int cx, cy;
